Add multeType

+++ src/main/java/com/teamManager/controller/MulteTypeController.java
@@ -1,5 +1,6 @@
 package com.teamManager.controller;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.teamManager.model.MulteType;
 import com.teamManager.repository.IMulteTypeRepository;
+import com.teamManager.service.TeamService;
 
 /**
  * The Class MulteTypeController.
@@ -24,6 +26,9 @@ public class MulteTypeController {
 
 	@Autowired
 	private IMulteTypeRepository multeTypeRepository;
+
+	@Autowired
+	private TeamService teamService;
 
 	/**
 	 * Adds the.
@@ -36,6 +41,7 @@ public class MulteTypeController {
 	 */
 	@RequestMapping(value = { "admin/multaType" }, method = RequestMethod.POST, consumes = "application/json")
 	public @ResponseBody Boolean add(@RequestBody MulteType multaType) throws Exception {
+		multaType.setTeam(teamService.getCurrentTeam());
 		multeTypeRepository.save(multaType);
 		return true;
 	}
@@ -51,23 +57,30 @@ public class MulteTypeController {
 	 */
 	@RequestMapping(value = { "admin/multaType" }, method = RequestMethod.PUT, consumes = "application/json")
 	public @ResponseBody Boolean update(@RequestBody MulteType multaType) throws Exception {
+		multaType.setTeam(teamService.getCurrentTeam());
 		multeTypeRepository.save(multaType);
 		return true;
 	}
 
-	/**
-	 * Gets the.
-	 *
-	 * @param id
-	 *            the id
-	 * @return the optional
-	 * @throws Exception
-	 *             the exception
-	 */
-	@RequestMapping(value = { "multaType" }, method = RequestMethod.GET, produces = { "application/json" })
-	public @ResponseBody Optional<MulteType> get(@NonNull @RequestParam Long id) throws Exception {
-		return multeTypeRepository.findById(id);
-	}
+	// /**
+	// * Gets the.
+	// *
+	// * @param id
+	// * the id
+	// * @return the optional
+	// * @throws Exception
+	// * the exception
+	// */
+	// @RequestMapping(value = { "multaType" }, method = RequestMethod.GET,
+	// produces = { "application/json" })
+	// public @ResponseBody Optional<MulteType> get(@NonNull @RequestParam Long
+	// id) throws Exception {
+	// Optional<MulteType> findById = multeTypeRepository.findById(id);
+	// if(findById != null){
+	// Team checkTeam = teamService.checkTeam(findById.get().getTeam());
+	// }
+	// return multeTypeRepository.findById(id);
+	// }
 
 	/**
 	 * Gets the all.
@@ -78,7 +91,7 @@ public class MulteTypeController {
 	 */
 	@RequestMapping(value = { "multaType/all" }, method = RequestMethod.GET, produces = { "application/json" })
 	public @ResponseBody Iterable<MulteType> getAll() throws Exception {
-		return multeTypeRepository.findAll();
+		return teamService.getCurrentTeam().getMulteTypes();
 	}
 
 	/**
@@ -87,10 +100,15 @@ public class MulteTypeController {
 	 * @param id
 	 *            the id
 	 * @return the boolean
+	 * @throws Exception
 	 */
 	@RequestMapping(value = { "admin/multaType" }, method = RequestMethod.DELETE, consumes = "application/json")
-	public @ResponseBody Boolean delete(@NonNull @RequestParam Long id) {
-		multeTypeRepository.deleteById(id);
+	public @ResponseBody Boolean delete(@NonNull @RequestParam Long id) throws Exception {
+		List<MulteType> multeTypes = teamService.getCurrentTeam().getMulteTypes();
+		Optional<MulteType> findById = multeTypeRepository.findById(id);
+		if (findById.isPresent() && multeTypes.contains(findById.get())) {
+			multeTypeRepository.deleteById(id);
+		}
 		return true;
 	}
 }
